исправил ошибку с получением ширины экрана, меню пользователя теперь не скрыто за блоком промо на главной стр
